Fix delete event endpoint

+++ GamesCorner/Server/Handlers/DeleteEventHandler.cs
@@ -8,11 +8,11 @@ public class DeleteEventHandler : IRequestHandler<DeleteEventRequest, IResult>
 {
 	public async Task<IResult> Handle(DeleteEventRequest request, CancellationToken cancellationToken)
 	{
-		//if (!request.HttpContextAccessor.HttpContext.User.IsInRole("Administrator") &&
-		//    !request.AuthService.ValidateToken(request.Token))
-		//{
-		//	return Results.Unauthorized();
-		//}
+		if (!request.HttpContextAccessor.HttpContext.User.IsInRole("Administrator") &&
+			!request.AuthService.ValidateToken(request.Token))
+		{
+			return Results.Unauthorized();
+		}
 		await request.UnitOfWork.EventRepository.DeleteByIdAsync(request.EventId);
 		await request.UnitOfWork.Save();
 		return Results.Ok("Event deleted");
